Added options menu

+++ include/2048util.h
@@ -7,4 +7,4 @@
 
 #include <macros.h>
 
-bool LoadIniFile(const char* path, glm::vec2* boardSize);
+bool LoadIniFile(const char* path, glm::ivec2* boardSize);
